Return to noarch build (#13)
* Return to noarch build

We used different Python version builds because importlib-metadata is
only required for Python 3.7 and lower. However, the advantages of
noarch are big enough that it's worth "unnecessarily" depending on
importlib-metadata in newer Python in order to avoid arch-specific
builds. (e.g. there are no native Apple Silicon builds of
napari-plugin-engine at the moment.) Once Python 3.7 is dropped from
conda-forge, we can drop the dependency.

* Add minimum Python version to recipe